BUGFIX: deploy mysql load no longer reports success erroneously.
There was a ; between the mysql commands and the rm command that it runs, which meant that a non-zero return code from the mysql command got swallowed. This has been corrected by using && instead, and the code that deals with those return values has also been made a bit more robust.